requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
mistralai>=0.0.10
colorama>=0.4.6 